Extract runner and runners from ParallelNaverClient
